fix(secretsmanager): Secret.fromSecretName doesn't work with ECS
The ability to import and reference a Secret purely by the secret name was
introduced in #10309. One of the original requests was modelled after the
integration with CodeBuild, where either the secret name or the full ARN
-- including the SecretsManager-provided suffix -- were accepted, but not a
"partial" ARN without the suffix. To ease integrations with other services
in this case, the `secretArn` was defined as returning the `secretName` for
these secrets imported by name.

However, other services -- like ECS -- require that an ARN format is provided,
even as a partial ARN. This introduces a dual behavior where sometimes the
secretName works and partial ARN fails, and other times the partial ARN works
and the secretName fails.

This change introduces an option to the `fromSecretName` factory to control this
behavior, so users can set up the secret properly for the service they are
integrating with.

*Disclaimer:* - I don't *love* this, and am very open to feedback on alternative
approaches that would also be backwards compatible.

Related changes -- I improved the suffix-strippiung logic of `parseSecretName`
to only strip a suffix if it's exactly 6 characters long, as all SecretsManager
suffixes are 6 characters. This prevents accidentally stripping the last word
off of a hyphenated secret name like 'github-token'.

fixes #10519